+++ src/parser.js
@@ -14,6 +14,11 @@ import Ast from './ast';
 import DataStructureGenerator from './schema';
 
 const FORM_CONTENT_TYPE = 'application/x-www-form-urlencoded';
+// Content types capable of being produced from formData
+const FORM_DATA_CONTENT_TYPES = [
+  'application/x-www-form-urlencoded',
+  'multipart/form-data',
+];
 
 // Provide a `nextTick` function that either is Node's nextTick or a fallback
 // for browsers
@@ -28,6 +33,15 @@ function nextTick(cb) {
 // Test whether a key is a special Swagger extension.
 function isExtension(value, key) {
   return _.startsWith(key, 'x-');
+}
+
+function isValidContentType(contentType) {
+  try {
+    typer.parse(contentType);
+  } catch (e) {
+    return false;
+  }
+  return true;
 }
 
 function isJsonContentType(contentType) {
@@ -800,56 +814,91 @@ export default class Parser {
     });
   }
 
+  // Returns all of the content types for a request
+  // Request content types include all consumes types
+  // Returns `[null]` when there are no content types
+  gatherRequestContentTypes(methodValue) {
+    const contentTypes = (methodValue.consumes || this.swagger.consumes || [])
+      .filter(isValidContentType);
+
+    if (contentTypes.length === 0) {
+      return [null];
+    }
+
+    return contentTypes;
+  }
+
+  // Returns all of the content types for a response
+  // Response content types include all produces types including any examples
+  // Returns `[null]` when there are no content types
+  gatherResponseContentTypes(methodValue, examples) {
+    let contentTypes = (methodValue.produces || this.swagger.produces || []);
+
+    // Add any missing produces listed in examples
+    if (examples) {
+      const exampleContentTypes = Object.keys(examples)
+        .filter(example => !contentTypes.includes(example));
+      contentTypes = contentTypes.concat(exampleContentTypes);
+    }
+
+    contentTypes = contentTypes.filter(isValidContentType);
+
+    if (contentTypes.length === 0) {
+      return [null];
+    }
+
+    return contentTypes;
+  }
+
   // Convert a Swagger response & status code into Refract transactions.
   handleSwaggerResponse(
     transition, method, methodValue, transitionParams,
     responseValue, statusCode, schemes, resourceParams,
   ) {
-    let examples;
+    const requestContentTypes = this.gatherRequestContentTypes(methodValue);
+    const responseContentTypes = this.gatherResponseContentTypes(methodValue,
+                                                                 responseValue.examples);
 
-    if (responseValue.examples) {
-      examples = responseValue.examples;
-    } else {
-      // The only way to specify an HTTP method is by creating a transaction,
-      // and according to the Refract spec a transaction *MUST* have a request
-      // and response within it, so here we seed the examples to create a blank
-      // request/response within a new transaction. See:
-      // https://github.com/refractproject/refract-spec/blob/master/namespaces/api-description-namespace.md#http-transaction-element
-      examples = {
-        '': undefined,
-      };
-    }
+    responseContentTypes.forEach((responseContentType) => {
+      let responseBody;
 
-    examples = _.omit(examples, 'schema');
+      if (responseContentType && responseValue.examples &&
+          responseValue.examples[responseContentType]) {
+        responseBody = responseValue.examples[responseContentType];
+      }
 
-    _.forEach(examples, (responseBody, contentType) => {
-      const transaction = this.createTransaction(transition, method, schemes);
+      requestContentTypes.forEach((requestContentType) => {
+        const transaction = this.createTransaction(transition, method, schemes);
 
-      this.handleSwaggerExampleRequest(transaction, methodValue, transitionParams, resourceParams);
-      this.handleSwaggerExampleResponse(transaction, methodValue, responseValue,
-                                        statusCode, responseBody, contentType);
+        this.handleSwaggerExampleRequest(transaction, methodValue, transitionParams,
+                                         resourceParams, requestContentType, responseContentType,
+                                         responseBody === undefined);
+        this.handleSwaggerExampleResponse(transaction, methodValue, responseValue,
+                                          statusCode, responseBody, responseContentType);
+      });
     });
   }
 
   // Convert a Swagger example into a Refract request.
-  handleSwaggerExampleRequest(transaction, methodValue, transitionParams, resourceParams) {
+  handleSwaggerExampleRequest(
+    transaction, methodValue, transitionParams, resourceParams,
+    contentType, responseContentType, contentTypeFromProduces,
+  ) {
     const request = transaction.request;
 
     this.withPath(() => {
-      // Check if json is in consumes
-      const consumes = methodValue.consumes || this.swagger.consumes || [];
-      const produces = methodValue.produces || this.swagger.produces || [];
+      const consumeIsJson = contentType && isJsonContentType(contentType);
 
-      const jsonConsumesContentType = _.find(consumes, isJsonContentType);
-      const jsonProducesContentType = _.find(produces, isJsonContentType);
-
-      // Add content-type headers
-      if (jsonConsumesContentType) {
-        pushHeader('Content-Type', jsonConsumesContentType, request, this, 'consumes-content-type');
+      if (contentType) {
+        pushHeader('Content-Type', contentType, request, this, 'consumes-content-type');
       }
 
-      if (jsonProducesContentType) {
-        pushHeader('Accept', jsonProducesContentType, request, this, 'produces-accept');
+      if (responseContentType) {
+        if (contentTypeFromProduces) {
+          pushHeader('Accept', responseContentType, request, this, 'produces-accept');
+        } else {
+          pushHeader('Accept', responseContentType, request, this);
+        }
       }
 
       const formParams = [];
@@ -879,8 +928,8 @@ export default class Parser {
                 }
 
                 this.withPath('schema', () => {
-                  if (jsonConsumesContentType) {
-                    bodyFromSchema(param.schema, request, this, jsonConsumesContentType);
+                  if (consumeIsJson) {
+                    bodyFromSchema(param.schema, request, this, contentType);
                   }
 
                   this.pushSchemaAsset(param.schema, request, this.path);
@@ -908,7 +957,9 @@ export default class Parser {
         });
       });
 
-      this.generateFormParameters(formParams, formParamsSchema, request);
+      if (!contentType || FORM_DATA_CONTENT_TYPES.includes(contentType)) {
+        this.generateFormParameters(formParams, formParamsSchema, request, contentType);
+      }
 
       // Using form parameters instead of body? We will convert those to
       // data structures and will generate form-urlencoded body.
@@ -916,16 +967,23 @@ export default class Parser {
     });
   }
 
-  generateFormParameters(parameters, schema, request) {
+  generateFormParameters(parameters, schema, request, contentType) {
     if (_.isEmpty(parameters)) {
       return;
     }
 
     const { DataStructure, Object: ObjectElement } = this.minim.elements;
 
-    pushHeader('Content-Type', FORM_CONTENT_TYPE, request, this, 'form-data-content-type');
+    if (!contentType) {
+      // No content type was provided, lets default to first form
+      pushHeader('Content-Type', FORM_CONTENT_TYPE, request, this, 'form-data-content-type');
+    }
 
-    bodyFromSchema(schema, request, this, FORM_CONTENT_TYPE);
+    if (!contentType || contentType === FORM_CONTENT_TYPE) {
+      // Only produce a body for FORM_CONTENT_TYPE (default when content type is null)
+      // We do not want to produce these bodies for other types such as multipart
+      bodyFromSchema(schema, request, this, contentType || FORM_CONTENT_TYPE);
+    }
 
     // Generating data structure
     const dataStructure = new DataStructure();
@@ -979,17 +1037,16 @@ export default class Parser {
       }
 
       if (contentType) {
-        this.withPath('examples', contentType, () => {
-          pushHeader('Content-Type', contentType, response, this);
-        });
+        if (responseValue.examples && responseValue.examples[contentType]) {
+          this.withPath('examples', contentType, () => {
+            pushHeader('Content-Type', contentType, response, this);
+          });
+        } else {
+          pushHeader('Content-Type', contentType, response, this, 'produces-content-type');
+        }
       }
 
-      const produces = methodValue.produces || this.swagger.produces || [];
-      const jsonProducesContentType = _.find(produces, isJsonContentType);
-
-      if (jsonProducesContentType) {
-        pushHeader('Content-Type', jsonProducesContentType, response, this, 'produces-content-type');
-      }
+      const isJsonResponse = isJsonContentType(contentType);
 
       if (responseValue.headers) {
         this.updateHeaders(response, responseValue.headers);
@@ -1038,8 +1095,8 @@ export default class Parser {
           }
 
           this.withSlicedPath(...args.concat([() => {
-            if (jsonProducesContentType !== undefined && responseBody === undefined) {
-              bodyFromSchema(schema, response, this, jsonProducesContentType);
+            if (isJsonResponse && responseBody === undefined) {
+              bodyFromSchema(schema, response, this, contentType);
             }
 
             this.pushSchemaAsset(schema, response, this.path);